+++ src/components/popover/popover.tsx
@@ -12,14 +12,7 @@ import {
   h,
   VNode
 } from "@stencil/core";
-import {
-  CSS,
-  ARIA_CONTROLS,
-  ARIA_EXPANDED,
-  HEADING_LEVEL,
-  TEXT,
-  defaultPopoverPlacement
-} from "./resources";
+import { CSS, ARIA_CONTROLS, ARIA_EXPANDED, TEXT, defaultPopoverPlacement } from "./resources";
 import {
   FloatingCSS,
   OverlayPositioning,
@@ -537,7 +530,7 @@ export class Popover implements FloatingUIComponent, OpenCloseComponent, FocusTr
   renderHeader(): VNode {
     const { heading, headingLevel } = this;
     const headingNode = heading ? (
-      <Heading class={CSS.heading} level={headingLevel || HEADING_LEVEL}>
+      <Heading class={CSS.heading} level={headingLevel}>
         {heading}
       </Heading>
     ) : null;