+++ src/components/value-list/value-list.tsx
@@ -22,6 +22,8 @@ import {
   deselectRemovedItems,
   getItemData,
   handleFilter,
+  handleFilterEvent,
+  handleInitialFilter,
   initialize,
   initializeObserver,
   ItemData,
@@ -68,6 +70,20 @@ export class ValueList<
   @Prop({ reflect: true }) dragEnabled = false;
 
   /**
+   * **read-only** The currently filtered items
+   *
+   * @readonly
+   */
+  @Prop({ mutable: true }) filteredItems: HTMLCalciteValueListItemElement[] = [];
+
+  /**
+   * **read-only** The currently filtered items
+   *
+   * @readonly
+   */
+  @Prop({ mutable: true }) filteredData: ItemData = [];
+
+  /**
    * When `true`, an input appears at the top of the component that can be used by end users to filter list items.
    */
   @Prop({ reflect: true }) filterEnabled = false;
@@ -76,6 +92,11 @@ export class ValueList<
    * Placeholder text for the filter's input field.
    */
   @Prop({ reflect: true }) filterPlaceholder: string;
+
+  /**
+   * Text for the filter input field.
+   */
+  @Prop({ reflect: true, mutable: true }) filterText: string;
 
   /**
    * The component's group identifier.
@@ -152,6 +173,8 @@ export class ValueList<
 
   emitCalciteListChange: () => void;
 
+  emitCalciteListFilter: () => void;
+
   filterEl: HTMLCalciteFilterElement;
 
   assistiveTextEl: HTMLSpanElement;
@@ -169,6 +192,7 @@ export class ValueList<
 
   componentDidLoad(): void {
     this.setUpDragAndDrop();
+    handleInitialFilter.call(this);
   }
 
   componentDidRender(): void {
@@ -197,6 +221,11 @@ export class ValueList<
    * Emits when the order of the list has changed.
    */
   @Event({ cancelable: false }) calciteListOrderChange: EventEmitter<any[]>;
+
+  /**
+   * Emits when a filter has changed.
+   */
+  @Event({ cancelable: false }) calciteListFilter: EventEmitter<void>;
 
   @Listen("focusout")
   calciteListFocusOutHandler(event: FocusEvent): void {
@@ -249,6 +278,10 @@ export class ValueList<
     this.filterEl = el;
   };
 
+  setFilteredItems = (filteredItems: HTMLCalciteValueListItemElement[]): void => {
+    this.filteredItems = filteredItems;
+  };
+
   setUpDragAndDrop(): void {
     this.cleanUpDragAndDrop();
 
@@ -281,6 +314,8 @@ export class ValueList<
   selectSiblings = selectSiblings.bind(this);
 
   handleFilter = handleFilter.bind(this);
+
+  handleFilterEvent = handleFilterEvent.bind(this);
 
   getItemData = getItemData.bind(this);